fix: cpf type to bigint

+++ priv/repo/migrations/20220119171236_initial.exs
@@ -4,7 +4,7 @@ defmodule Cumbuca.Repo.Migrations.Initial do
   def change do
     create table("user") do
       add :name, :string
-      add :cpf, :integer
+      add :cpf, :bigint
       add :email, :string
       add :balance, :integer
